+++ package.json
@@ -24,24 +24,30 @@
     "apollo-client": "^2.3.5",
     "apollo-link": "^1.2.2",
     "apollo-link-http": "^1.5.4",
+    "apollo-server": "^2.0.0-rc.4",
     "axios": "^0.18.0",
     "body-parser": "^1.18.3",
+    "cors": "^2.8.4",
     "dotenv": "^6.0.0",
     "express": "^4.16.3",
-    "express-graphql": "^0.6.12",
+    "express-jwt": "^5.3.1",
     "express-session": "^1.15.6",
     "graphql": "^0.13.2",
     "graphql-tag": "^2.9.2",
+    "graphql-tools": "3.0.2",
     "helmet": "^3.12.1",
     "lodash": "^4.17.10",
     "morgan": "^1.9.0",
     "mysql2": "^1.5.3",
+    "jsonwebtoken": "^8.3.0",
     "react-apollo": "^2.1.6",
     "react-router-dom": "^4.3.1",
     "sequelize": "^4.38.0"
   },
   "devDependencies": {
     "babel-core": "^6.26.3",
+    "babel-inline-import-loader": "^0.1.0",
+    "babel-plugin-inline-import": "^3.0.0",
     "babel-loader": "^7.1.4",
     "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.24.1",
@@ -49,6 +55,7 @@
     "html-loader": "^0.5.5",
     "html-webpack-plugin": "^3.2.0",
     "mini-css-extract-plugin": "^0.4.0",
+    "nodemon": "^1.17.5",
     "react": "^16.4.1",
     "react-dom": "^16.4.1",
     "webpack": "^4.12.0",
