--- server/middleware/schema.js
@@ -1,3 +1,0 @@
-const graphql = require('graphql');
-const {GraphQLSchema} = graphql;
-
