+++ server/middlewares/schema.graphql
@@ -1,0 +1,32 @@
+
+type Query {
+  # company(id: ID!): Company
+  # job(id: ID!): Job
+  jobs: [Job]
+}
+
+# type Mutation {
+#   createJob(input: CreateJobInput): Job
+# }
+
+# type Company {
+#   id: ID!
+#   name: String
+#   description: String
+#   jobs: [Job]
+# }
+
+# Create a query template for Job
+# We're telling graphQL whenever we ask for Job,
+# give us the follow fields.
+# ! mark means the field is not null.
+# graphQL is statically typed, by default there are
+# ID, Int, String, Boolean, Float.  
+type Job {
+  id: ID!
+  description: String
+  subject: String
+  grade: Int
+}
+
+
